Transition API (#169)

+++ compositor_common/src/util/degree.rs
@@ -1,4 +1,4 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Clone, Default, PartialEq, Eq)]
-pub struct Degree(pub i32);
+#[derive(Debug, Serialize, Deserialize, Clone, Default, PartialEq)]
+pub struct Degree(pub f64);
